First checkins of the DojoX A/V project.  This checkin only includes the base files needed to include Flash and QuickTime movies on the fly, more to follow.
A few notes:
Eventually the base objects here will replace dojox.flash; given the current situation (i.e. common versions of Flash available), we're going to remove support for Flash 6 and go with a minimum of Flash 8 (as supported by MTASC).  This means the cross-version ExternalInterface code in dojox.flash can be seriously paired back or eliminated altogether.

Parts of dojox.flash will be incorporated into DojoX A/V (specifically, some of the chunking methods and the ExpressInstall class handling), with an aim of eventually removing dojox.flash altogether (dojox.storage will eventually be modified to use DojoX A/V as the source for flash movies).

A final note: the only way around the Eolas hotfix (IE) is to place the script that actually pushes Flash movies into a document in an externally loaded script; since the most common form of the Dojo package system uses inline eval, there is special code in dojox/av/_base/flash.js to append an external script to the document.  This is only for IE; no other browsers are affected.