!strict fixes #6103 Grid resize logic changed to correct this behavior. The issue here is that the header width must be set before its height can be properly calculated.
Previously the header was assumed to be the correct height right after it was rendered in Grid.prerender. Then all dimensions were set for each view at once in Grid._resize.

Now, the header is rendered but not sized. Then view widths and position are set. Then header and view heights are set. This should correct the problem while maintaining efficient rendering.