refs #7018 - at first it seemed a regression in dojo.string.substitute, which fails if the value passed is undefined. i suspect this is a regression in FlickrRestStore, as it only affects the SlideShow's using that. onBegin seems to fire too often, and in the case of FlickrStore is not passing the count as it was before. Adding a workaround to slurp items.length onComplete and adding a sanity check to the maxPhotos var in the string.substitute() call in _setTitle fixes any errors seen, but this is definately a regression in one of those two places. !strict also minor style cleanups across the board.
also exposes some jankyness -- currently nightly displays same issue though. in IE7, the overall widget isn't being sized
if the image is in cache (it seems) but could be related to mis-firing of FlickrRest Store. Going to pass to Bryan to look
into FlickrRestStore, but pass back if it's an image issue vs a store-event-firing issue.